Move `PrerequisiteSet` to new file
